Refactor TransactionsRequest and init mappers

+++ src/main/java/com/EM_System/pojo/Account.java
@@ -24,7 +24,7 @@ public class Account implements CreateRequestItem {
         return balance;
     }
 
-    public void addBalance(int amount) {
+    public void addBalance(double amount) {
         this.balance += amount;
     }
